Implement decoding literal headers
Headers that are represented as literals can now also be decoded.

The dynamic table is also managed, given that the functionality
is encapsulated in the `DynamicTable` struct, which is used by the
`Decoder`.

However, only literals that are not Huffman encoded can be read,
since decoding Huffman octet strings is not implemented yet.