Create a Huffman decoder from a table of codes
The Huffman decoder reads in the given table of codes and constructs
a simple table mapping the code length to a mapping of all codes of
that length to their original symbol.

This way it is possible to unambigously decode any codepoint defined
by a (right-aligned `u32` representation of the code, number of bits
in the code) tuple.