HuffmanDecoder: Implement the `decode` method
The decoding of a buffer containing the bytes representing a Huffman
encoded octet string is now implemented.

Error checking (such as invalid padding) is foregone for now.

The implementation goes through each individual bit in the buffer
(relying on the `BitIterator` struct), keeping track of the current
right-aligned code point and its length. Once such a codepoint can
be found in the table, we are sure it's a decoded symbol. At this
point the codepoint is reset and the process starts from the
beginning searching for the next valid codepoint.