Align some code to column width 80
The test also contained a byte literal that exceeded the value 255.
However, this was not an issue as it was there simply to add padding
to the buffer to test how the decoder handles having more bytes than
needed. Regardless, the value is now brought down to the correct
range.