Refactor Decoder to hold a borrow of a static table
So far, the Decoder was directly referencing the global/static
STATIC_TABLE. Now that the static table has been moved to the
parent module, it's a bit of a code smell that the implementation
stil referencing this global all over the place.

The solution is a new constructor type that lets us construct
a `Decoder` with a given static table. The default constructor
(i.e. Decoder::new) now simply passes the `STATIC_TABLE` to this
constructor, keeping the old behavior, but cleaning up the code.