Implement an Iter through the entire HeaderTable
This iterator makes it possible to conveniently go through each
header found in the entire index address space of headers, without
explicitly referring to the static or the dynamic part of the
address space.

The implementation required some **very** ugly type declarations
and workarounds for the fact that
["abstract return types"](https://github.com/rust-lang/rfcs/pull/105)
don't exist yet -- i.e. it is not possible to have the compiler
infer the exact return type of the iterator obtained by chaining two
iterators (and an odd `map` in the mix). However, the `iter` method
now provides a clean interface for users to iterate through the
table and this ugliness is contained to a single struct, making it
a valuable trade-off.