HeaderTable: Implement a lookup method for headers
The method makes it possible to check whether a given header is
found in the header table. The method tries to match both the name
and the value, if possible. If no such header is found, it returns
one that has the same header name. Finally, if that is not found
either, returns `None`.