Add an HPACK Encoder struct
The struct will serve as the main API for accessing HPACK header
encoding functionality of the `solicit::hpack` module.

A new `Encoder` is created with a default context: an empty dynamic
table (of default size) and a static table corresponding to the
one defined in the Appendix A of the HPACK spec.