Encoder: Implement initial HPACK encoding
The initial implementation is extremely simple, but follows all
HPACK rules so that it can be successfully decoded by compliant
decoders.

In particular, no Huffman string encoding is supported and the
algorithm for choosing which header representation is used and
when headers are added to the dynamic table is also very simple:
any header already in the table is always encoded using the
index; headers where the name was not previously in the table
are added to the table (incremental indexing); headers where only
the name was found, simply get encoded with an indexed name
(and not added to the table).