Add a RawFrame struct: unparsed HTTP/2 frames
The `RawFrame` struct simply stores the raw components of an
HTTP/2 frame: its header and the payload as a sequence of bytes.
It does not try to interpret the bytes, nor do any validation in
terms of its validity based on the frame type given in the header.
It is simply a wrapper around the two parts of an HTTP/2 frame.