Implement reading frames from an `HttpConnection`
Frames are read from the underlying `TransportSocket`, decoded into
an appropriate `Frame` implementation and returned wrapped in a
corresponding `HttpFrame` variant.

Any IO errors are propagated.

The connection does not give exact reasons why a frame was deemed
invalid yet, nor does it take any action when an invalid frame is
encountered (such as sending a PROTOCOL_ERROR to the peer).