Add a Stream trait for representing HTTP/2 streams
The trait defines which operations need to be defined by a type that should
be useable as an HTTP/2 stream. By implementing this trait, clients can only
implement stream-level logic, such as how the received data should be handled,
instead of tracking which streams exist and what their states are.