Implement a simple HTTP/2 client
This client allows the user to issue a number of requests before
blocking the thread by waiting for a response to one of those requests.

Once that request returns with a response, the client can issue more
requests before waiting for another response.

This is similar to what an HTTP/1.1 client might do with a keep-alive
connection with request pipelining. In this case, however, the
responses for the issued requests do not have to come in the order
that they were issued, but this client does not expose any way to
notify the user which request got a response; it needs that the user
asks for a particular request's response to keep similarity with an
HTTP/1.1 client.