Simplify converting io::Errors to HttpErrors
The `FromError` trait allows for easy conversion from a particular
error into a different one. Additionally, the `try` macro already
uses it to convert the resulting error to the expected type.

Therefore, by implementing the `FromError<io:Error>` for the `HttpError`
we can get rid of the custom `try_io` macro that was wrapping the
`io::Error` into an `IoError` variant of the `HttpError`.