Implement a TLS connector
This connector establishes an HTTP/2 connection over a TLS protected
socket and negotiates the use of HTTP/2 during the TLS handshake.

It provides the option to pass a custom SSL context, but still
validates that the final chosen protocol is HTTP/2, returning an error
if not. The error includes the socket, so clients can decide whether
to throw it away or to pass it to a different component that can speak
the protocol that *was* chosen.