Refactor the frame type extraction to HttpFrame
Deciding the frame type based on the header was being done in the
`recv_frame` method of `HttpConnection`. This, however, means that the
logic of which IDs correspond to which type is untestable separately
from the `HttpConnection` itself, as well as not reusable. This is now
extracted to a static method on the `HttpFrame` type, achieving both of
the previously stated goals.