Stop treating unknown frames as Http errors
Instead of returning an error variant from `recv_frame` when the frame
type is unknown, now we wrap the `RawFrame` into an
`HttpFrame::UnknownFrame` variant and return that, possibly allowing the
session layer in the future to provide handling for unknown frames too.

The `UnknownFrameType` variant of the `HttpError` is removed, as it is
now obsolete.