Extract the host name into the ClientConnection
This is because if we want to make the `HttpConnection` itself work
regardless of whether it is a server-side or a client-side connection,
it cannot have a hostname field, given that that's associated only with
client-side connections (i.e. the server does not need to know the
domain name of its peer).