Shift the preface-writing responsibility out of the HttpConnection
The responsibility is now on the code that creates the underlying
stream to provide it to the `HttpConnection` in a correct state. The
`HttpConnection` wants to deal only with HTTP/2 frames, not any sort of
raw bytes.

The helper `write_preface` method writes the client preface to any
`io::Write`. Additionally, the `HttpConnect` implementations are adapted
to automatically write the headers so that any HTTP/2 implementation
based on native Rust IO does not have to worry about this detail at all.