Repurpose the Connector API: Establish Transport connections
Instead of creating full `HttpConnection` instances, the `HttpConnect`
API now requires that a `TransportStream` is created instead. This makes
this API a helpful shortcut to establishing `TransportStream` instances
that are compliant with what HTTP/2 connections require (in terms of the
settings and initialization).

This is a step towards making `HttpConnections` independent of the
specifics of the underlying IO.