HttpConnection: Use a separate `ReceiveFrame` instance
The stream given to the connection is now split into two parts:
one is used as a `ReceiveFrame` instance and the other as `SendFrame`.

This is work in progress in order to disassociate `HttpConnection`s from
particular IO implementations and have them only depend on having
facilities for obtaining/writing HTTP/2 frames.