HttpConnection tests: Make an unsafe block safe
The block transmutes a slice into a fixed-length array reference,
however it wasn't asserting that the slice is large enough for the fixed
length array, which could lead to a segfault if proven incorrect.

The block now `assert!`s this before performing the transmute, making
the process panic if the assumption is false.