Move incoming frame handling to HttpConnection
The `ClientConnection` contained the logic of parsing incoming frames
and dispatching events to a `Session`. However, this is a common
functionality required by both client- as well as server-side
connections. Therefore, the functionality is simply factored up into the
`HttpConnection`, with the `ClientConnection` now only providing the
session into which the events are pumped by the `HttpConnection`'s
`handle_next_frame` method.