Remove references to the host name from HTTP connections
The HTTP connection itself does not require knowledge of the host name
-- and indeed cannot even sensibly contain one for server-side
connections. This commit removes the host fields from the
`HttpConnection` and shifts it all the way up to the client
implementations to insert the host name into the ":authority" header
before issuing a request by delegating to the `HttpConnection` to send
the header list.

This is the final step toward making the `HttpConnection` fully
independent from whether it is used for a server- or a client-side
HTTP/2 connection!