HttpConnection tests: Refactor session integration tests
The tests that were asserting that the `HttpConnection` raises the
correct events on the `Session` instance that it is given were being
tested implicitly through the `ClientConnection` -- assuming that the
`ClientConnection` would pass its `Session` onto the `HttpConnection`.

These tests had too many moving parts and it is better to test the
corresponding `HttpConnection` methods directly.