Refactor(ClientConnection): Store the session state directly
The `ClientConnection` struct now stores the state of the connection
directly, instead of a `Session` implementation. Indeed, the purpose of
the `Session` implementation is to act as the interface between the
`HttpConnection` and the higher levels. `ClientConnection` is such a
higher level. Therefore, the `ClientConnection` now stores the state
directly in order to be able to inspect and manipulate it easier, while
providing a set of callbacks to the `HttpConnection` that can manipulate
this state as necessary.