ClientConnection: Add `start_request` helper method
The method starts a request based on a provided request stream by
immediately queuing the headers onto the HTTP/2 connection and adding
the given stream into the state. The given stream will handle the
response when it is received and provide the data that should be sent by
the `send_next_data` method.