client/async: Implement sending full request bodies
The async client now has an additional WorkItem that, when consumed,
triggers the `send_next_data` operation on the underlying connection.
This way, the client is able to queue more than a single data frame for
the request (as when `send_request` method was used).

For now, since there is no flow control to block sending the data, the
send data work item is queued after each frame is sent. This way, when a
new request is started (i.e. the headers frame sent), the data will
follow, if any. Additionally, any extra data frames will be queued due
to the send data work item that follows the send of the previous data.

In the future, when the sending of data can get blocked due to flow
control, the client will need to queue send operations after receiving
window update frames which unblock the stream.