server: Add a sample server implementation
This sime example of a server implementation allows for handling
requests arising on an HTTP/2 connection (only after fully reeiving the
request). It is only an example as the IO is not managed to allow for
concurrent reads and writes from the underlying socket (writing the
response to a single request blocks reads until the full body's been
flushed).