TransportStream: Make the `close` method require &mut self
Although the TCP/TLS-based connections do not require &mut, other
implementations might. This would introduce unnecessary strain when
trying to implement the `TransportStream` trait for those types (a
wrapper type that is based on some interior mutability sync mechanism
would be required). On the other hand, there don't seem to be any
advantages in holding on to a non-mutable method...