client/simple: Return an `Err` when a connector fails
Instead of panicking (on an `unwrap()`), the error returned by the
connector is now propagated up to the user. Since the method was already
returning a `Result` type, there was no reason to panic.

In order to make sure that all connectors return a valid error type,
the `HttpConnect` trait now mandates that the associated error type
needs to implement `HttpConnectError`; conversion from this trait to
`HttpError` is now also provided.

Closes #3