Fix name collision for `read_exact` on `TransportStream` trait
The `std::io::Read` trait is about to get its own implementation of the
`read_exact` method (rust-lang/rust#27585).

This causes a name resolution problem on the nightly version since
the compiler is unable to decide which implementation of the method
should be used. This commit fixes the problem by explicitly requesting
the `TransportStream` method to be used.

Closes #13