RawFrame: drop the `with_payload` method
The method was only being used by tests and a single place in the rest
of the library. The method is definitely not the best way to construct a
new `RawFrame` as it will perform extra allocations and copies behind
the scenes and as such keeping it in the public API would encourage bad
style. Therefore, this method is now removed.

A helper function that can be used by tests to craft new `RawFrame`s for
the purpose of testing is added to replace the calls previously made to
it by test code.