RawFrame: Add `parse` method
The method takes a slice and attempts to extract a RawFrame starting
from the beginning of the slice. If the parse is successful, returns a
RawFrame that borrows the appropriate content from the given buffer. If
a RawFrame cannot be extracted, returns None.

The method does not, however, attempt to do any validation based on the
semantic type of the frame. It merely asserts that the frame has a valid
HTTP/2 frame structure (9-octet header and a payload of the size
indicated by the header).