Do not require the HttpConnection to own the frame receiver
By removing the requirement, we are further decoupling the
`HttpConnection` for the underlying IO mechanism, as well as improving
the efficiency of processing frames in certain cases.

While the existing implementation allowed the IO to be abstracted to a
simple frame provider, it still required that the `HttpConnection`
itself always owns the provider. This leads to a situation where frame
provider cannot provide a frame backed by a borrowed buffer, but
requires that frames always be converted to an owned representation in
one way or another.

This is unnecessary, though, as the `HttpConnection` only needs to
inspect the next frame and never keep it around.

With this commit, we make it so that the client of the `HttpConnection`
needs to pass in a `&mut ReceiveFrame` to the `handle_next_frame`. This
finally allows the `ReceiveFrame` instance itself to have borrows
independent of the `HttpConnection`'s lifetime.

The rest of the commit deals with the fallout of breaking this central
method's signature and adapts the session-layer clients bundled with
`solicit` to use it.