HttpConnection: Do not require ownership of the SendFrame instance
This commit changes the `HttpConnection` so that it no longer needs to
own the `SendFrame` instance that it uses to send frames, but that one
is provided by the client when they would like the HttpConnection to
craft frames to be sent.

This is achieved by providing a `sender` method on the `HttpConnection`
that returns an `HttpConnectionSender` instance that will use a given
`SendFrame` reference to queue crafted frames for sending.

Furthermore, the `HttpConnection` is now free of any generic type
parameters.

The commit deals with the fallout of drastically changing the send API,
as well as the loss of the generic type parameter...