Introduce a new `Header` struct
The new `Header` struct is able to represent headers without requiring
the name and value to be owned; indeed, the name and value can be either
borrowed from a buffer or owned by the `Header`. This allows for a more
efficient representation of headers across the board, but especially
when the header can be represented by two static buffers, such as is the
case for usual pseudo-headers (:status, :method) and their common values
(200, GET, etc.).