client/async: Refactor the frame reader
The frame provider no longer depends on an inner generic `ReceiveFrame`,
but rather a `TransportStream`.

The implementation reads a frame off the TransportStream and sends its
RawFrame representation to the service thread, where the RawFrame is
finally parsed into an HttpFrame.

The reason behind this change is that HttpFrames should only be a parsed
representation of a frame used by an `HttpConnection` and requiring them
to be `Send` limits us to having only variants that fully own their
frame buffers (or holds static references), which is an unnecessary hit
to efficiency. This commit starts the process of allowing HttpFrames to
simply borrow an underlying RawFrame and provide an appropriate API
depending on the underlying frame type.