MockSendFrame: Keep RawFrames, not parsed ones
As parsed frames are intended to become a temporary interpretation of
RawFrames (and as such should simply borrow existing RawFrames), it will
become impossible to keep around parsed frames without its backing raw
frame. Therefore, the mock now stores the raw frame and expects the
assertions to deal with this fact.