Tie a parsed frame's lifetime to an existing RawFrame
This makes it possibe for parsed frames to be a wrapper around the
provided raw frame and give the clients an appropriate API depending on
the type of the RawFrame. The parsed Frame is no longer required to copy
parts of the original RawFrame that it needs to hold on to.