SendFrame: Send `FrameIR` instances, instead of `Frame`
With this change, the difference between types that represent parsed
frames and types that represent frames that are being built up for
sending becomes clearer. Before, the `Frame` trait encompassed both of
these use cases, which caused subpar performance in many aspects (too
many buffer allocations and copies). Additionally, the design is just
better when the different responsibilities are cleanly separated into
different traits.

The commit also deals with the fallout from breaking the `SendFrame`
trait API across the crate.