Remove serialize method from Frame trait
The Frame trait is supposed to represent a parsed frame. As such, the
serialize method is out of place on the trait itself, even if the same
struct can represent both a parsed, as well as a serializable frame.

The method is now only conditionally compiled as a test helper for each
existing frame struct (this means that it is only ever used out of test
code).