HttpError: Add PeerConnectionError variant
The variant is to be used when the underlying error is caused by the
peer aborting the connection. When this is the case, we will have gotten
a GOAWAY frame that outlined the reason for aborting the connection (at
least an error code, optionally also some debug data).

This error can now be surfaced using the new variant.