Session: Add `on_goaway` method
The method is to be invoked by the HttpConnection when it detects that
the peer has reset he connection. It has a default implementation that
maps the error into an appropriate HttpError variant. This does not
preclude concrete Session implementations to provide a custom
implementation that also makes use of the last stream id...