HttpConnection: Notify session on window size decreases
Two new methods are added to the `Session` trait that allow it to
receive notice when either the connection-level or a stream-level
inbound window size decreases. These calls will, as a rule, always
be emitted in pairs by the `HttpConnection` upon parsing a flow-control
subject frame (at the moment this is only the DATA frame, as per the
HTTP/2 spec).