SessionState: Manage state Entries, not naked `Stream`s
The state now has an entry per stream, instead of storing just the raw
stream. This is in preparation for extending the state to be able to
track more stream-related information, that does not belong on the
Stream trait itself.

(The Stream trait is supposed to be the bridge between the session and
its client that is interested only in stream-level events; requiring
those clients to also implement tracking various book-keeping stream
state is not only redundant---all streams will need to track the same
state in the same way---but also not in line with what the Stream trait
is supposed to represent.)