Prevent invalid connections from panicing
If an invalid connection string was passed to an async
`Client::with_connector` then an unwrap error would occur. This error
would be a vague statement:

```
Option::unwrap()` on a `None` value'
```

This is rather hard to debug. When solicit is ready for breaking
changes, then this should be updated to return an HttpError, instead of
an Option.