Add WindowUpdateStrategy trait and an impl that disables flow control
The WindowUpdateStrategy trait is used to define the algorithm for
updating the flow control window of both the connection, as well as
individual stream.

The object is provided the new size of the corresponding window and
is expected to produce an action to be taken immediately: either
increase the size of the window or do nothing.