Remove unnecessary channel send in async client
When sending on this channel, it was assumed (safely, at this time) that
the send would succeed. It would be easy to violate this invariant
inadvertently. Rather than expose the possible SendError in the
ClientConnectError type, the send is obviated by calling the
ClientService.on_new_client method directly. This has the same effect as
sending a `WorkItem::NewClient` through the channel.