ci: Disable flaky live tests
Running the live tests against http2bin.org was meant as a top-level
integration test. However, the service has had temporary outages of late
and it's currently been down for a few days. Running these tests causes
CI to ostensibly fail, which is just a false positive in this case.

Disabling this kind of live test for now.

In the future, a more controlled fully locally controlled live test
could be used, instead for the same purpose. (For instance, perhaps
spawning a Python http/2 server locally and talking to that might be a
viable option.)